// generated/src/app/api-models/user-users.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class UserUsers {
  username: string;
  userId: number;
  accountStatus: string;
  lockDate: Date;
  expiryDate: Date;
  defaultTablespace: string;
  temporaryTablespace: string;
  localTempTablespace: string;
  created: Date;
  initialRsrcConsumerGroup: string;
  externalName: string;
  proxyOnlyConnect: string;
  common: string;
  oracleMaintained: string;
  inherited: string;
  defaultCollation: string;
  implicit: string;
  allShard: string;
}